\name{SubpathwayGMir-package}
\alias{SubpathwayGMir-package}
\alias{SubpathwayGMir}
\docType{package}
\title{The SubpathwayGMir package
}
\description{ SubpathwayGMir is an R package for identifying metabolic subpathways mediated by microRNAs (miRNAs).}
\section{Introduction}{
 SubpathwayGMir is an R package for identifying miRNA-mediated metabolic subpathways by topologically analyzing miRNAs and genes within reconstructed Kyoto Encyclopedia of Genes and Genomes (KEGG) pathway graphs, which integrated miRNA-target interactions verified by low-throughput experiments.
}
\author{ Li Feng, Chunquan Li and Xia Li}
\seealso{\code{\link{getInteGraphList}},\code{\link{getLocSubGraph}},\code{\link{identifyGraph}}, \code{\link{initializeK2ri}}, \code{\link{updateOrgEnvir}}}
